OXDEV-7733 Move Tracker repository to Repository directory
