+++ tests/Integration/Greeting/Controller/GreetingControllerTest.php
@@ -16,7 +16,7 @@ use OxidEsales\ModuleTemplate\Model\GreetingTracker;
 use OxidEsales\ModuleTemplate\Model\User as ModuleUser;
 use OxidEsales\ModuleTemplate\Service\ModuleSettingsServiceInterface;
 use OxidEsales\ModuleTemplate\Tests\Integration\IntegrationTestCase;
-use OxidEsales\ModuleTemplate\Tracker\Repository;
+use OxidEsales\ModuleTemplate\Tracker\Repository\TrackerRepository;
 
 /*
  * We want to test controller behavior going 'full way'.
@@ -52,7 +52,7 @@ final class GreetingControllerTest extends IntegrationTestCase
         $user->load(self::TEST_USER_ID);
         $this->assertSame($expected, $user->getPersonalGreeting());
 
-        $tracker = $this->get(Repository::class)
+        $tracker = $this->get(TrackerRepository::class)
             ->getTrackerByUserId(self::TEST_USER_ID);
         $this->assertSame($count, $tracker->getCount());
     }
